Fixing .travis.yml and adapting scripts

+++ script/install_lcov.sh
@@ -1,8 +1,16 @@
 #!/usr/bin/env bash
-echo "Installing lcov"
-cd $1
-wget http://ftp.de.debian.org/debian/pool/main/l/lcov/lcov_1.14.orig.tar.gz
-tar xf lcov_1.14.orig.tar.gz
-sudo make -C lcov-1.14/ install
-gem install coveralls-lcov
-rm lcov_1.14.orig.tar.gz
+echo "Checking folder ${1}lcov-1.14/"
+if test -f "${1}lcov-1.14/"; then
+    echo "Installing lcov"
+    cd ${1}
+    sudo make -C lcov-1.14/ install
+    gem install coveralls-lcov
+else
+    echo "Building and installing lcov"
+    cd ${1}
+    wget http://ftp.de.debian.org/debian/pool/main/l/lcov/lcov_1.14.orig.tar.gz
+    tar xf lcov_1.14.orig.tar.gz
+    sudo make -C lcov-1.14/ install
+    gem install coveralls-lcov
+    rm lcov_1.14.orig.tar.gz
+fi
